feat: Add persistence management for user and role configurations
